Simplify ParseTemporalMonthDayString
ParseTemporalMonthDayString could be simplfy for not dealing impossible condiction
https://tc39.es/proposal-temporal/#sec-temporal-parsetemporalmonthdaystring

Since we throw in step 2. If isoString does not satisfy the syntax of a TemporalMonthDayString (see 13.33), then
we know we have a TemporalMonthDayString which is either DateSpecMonthDay or DateTime
DateSpecMonthDay =  --opt DateMonth -opt DateDay
and
DateTime = Date TimeSpecSeparatoropt TimeZoneopt

and we know Date is either of the following two
DateYear - DateMonth - DateDay   
DateYear DateMonth DateDay

so in all three cases we have both DateMonth and DateDay
and if either DateMonth or DateDay is not presented then step 2 will throw.
Therefore, both 5. If month is undefined, and 7. If day is undefined,  can never happen